task 6

+++ 0x00-hello_world/6-size.c
@@ -15,10 +15,10 @@ int main(void)
 	long long lil;
 	float f;
 
-	printf("Size of a char: %ld byte(s)\n", sizeof(c));
-	printf("Size of an  int: %ld byte(s)\n", sizeof(i));
-	printf("Size of a long int: %ld byte(s)\n", sizeof(li));
-	printf("Size of a long long int: %ld byte(s)\n", sizeof(lil));
-	printf("Size of a float: %ld byte(s)\n", sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an  int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n",(unsigned long)sizeof(li));
+	printf("Size of a long long int: %lu byte(s)\n",(unsigned long)sizeof(lil));
+	printf("Size of a float: %lu byte(s)\n",(unsigned long)sizeof(f));
 	return (0);
 	}
